+++ app/components/bali/filters/preview.rb
@@ -3,7 +3,7 @@
 module Bali
   module Filters
     class Preview < ApplicationViewComponentPreview
-      FORM = DummyFilterForm.new
+      FORM = Bali::Utils::DummyFilterForm.new
 
       def default
         render Bali::Filters::Component.new(form: FORM, url: '#', text_field: :name)
@@ -14,7 +14,8 @@ module Bali
           c.attribute(
             title: 'Active',
             attribute: :status_in,
-            collection_options: [['active', true], ['inactive', false]])
+            collection_options: [['active', true], ['inactive', false]]
+          )
         end
       end
     end
