+++ spec/components/bali/data_table_spec.rb
@@ -3,7 +3,8 @@
 require 'rails_helper'
 
 RSpec.describe Bali::DataTable::Component, type: :component do
-  let(:component) { Bali::DataTable::Component.new(filter_form: DummyFilterForm.new, url: '#') }
+  let(:form) { Bali::Utils::DummyFilterForm.new }
+  let(:component) { Bali::DataTable::Component.new(filter_form: form, url: '#') }
 
   subject { rendered_component }
